Don't update overclock on the first boot

+++ packages/351elec/sources/autostart.sh
@@ -15,20 +15,23 @@ mkdir /tmp/logs
 # Apply some kernel tuning
 sysctl vm.swappiness=1
 
-# Restore overclock setting
-OVERCLOCK_SETTING=$(get_ee_setting overclock)
-OVERCLOCK_STATE=$((grep "\-oc.dtb" /flash/boot.ini >/dev/null 2>&1 && echo 1) || echo 0)
-if [ ! "${OVERCLOCK_STATE}" == "${OVERCLOCK_SETTING}" ]
+if [ -e "/storage/.newcfg" ]
 then
-  echo -en '\e[0;0H\e[37mRestoring overclock...\e[0m' >/dev/console
-  if [ "${OVERCLOCK_SETTING}" = "1" ]
+  # Restore overclock setting
+  OVERCLOCK_SETTING=$(get_ee_setting overclock)
+  OVERCLOCK_STATE=$((grep "\-oc.dtb" /flash/boot.ini >/dev/null 2>&1 && echo 1) || echo 0)
+  if [ ! "${OVERCLOCK_STATE}" == "${OVERCLOCK_SETTING}" ]
   then
-    /usr/bin/351elec-overclock on
-  else
-    /usr/bin/351elec-overclock off
+    echo -en '\e[0;0H\e[37mRestoring overclock...\e[0m' >/dev/console
+    if [ "${OVERCLOCK_SETTING}" = "1" ]
+    then
+      /usr/bin/351elec-overclock on
+    else
+      /usr/bin/351elec-overclock off
+    fi
+    sleep 1
+    systemctl reboot
   fi
-  sleep 1
-  systemctl reboot
 fi
 
 # Restore config if backup exists
